Move TrackSelectionList feature to list package
