+++ shared/src/androidMain/kotlin/org/hyperskill/app/track_selection/list/injection/PlatformTrackSelectionListComponent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.track_selection.injection
+package org.hyperskill.app.track_selection.list.injection
 
 import org.hyperskill.app.core.injection.ReduxViewModelFactory
