+++ shared/src/androidMain/kotlin/org/hyperskill/app/track_selection/list/injection/PlatformTrackSelectionListComponentImpl.kt
@@ -1,8 +1,7 @@
-package org.hyperskill.app.track_selection.injection
+package org.hyperskill.app.track_selection.list.injection
 
 import org.hyperskill.app.core.injection.ReduxViewModelFactory
-import org.hyperskill.app.track_selection.list.injection.TrackSelectionListComponent
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListViewModel
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListViewModel
 import ru.nobird.app.presentation.redux.container.wrapWithViewContainer
 
 class PlatformTrackSelectionListComponentImpl(
